chore: refactor CheckResult into a common class
fix: logic bug where success was never set to true
