+++ invenio_checks/contrib/file_formats.py
@@ -15,7 +15,7 @@ from pathlib import Path
 import yaml
 from flask import current_app
 
-from invenio_checks.base import Check
+from invenio_checks.base import Check, CheckResult
 from invenio_checks.models import CheckConfig
 from invenio_checks.utils import classproperty
 
@@ -60,21 +60,6 @@ class FileFormatDatabase(dict):
             for ext in ff_spec.extensions:
                 res._ext_lookup[ext].add(ff_spec.id)
         return res
-
-
-@dataclass
-class CheckResult:
-    """Result of a check."""
-
-    id: str
-    title: str
-    description: str
-    errors: list[dict] = field(default_factory=list)
-    success: bool = True
-
-    def to_dict(self):
-        """Convert the result to a dictionary."""
-        return asdict(self)
 
 
 class FileFormatsCheck(Check):
